move context in to seperate file
