+++ packages/react-core/src/components/Page/Page.tsx
@@ -8,36 +8,12 @@ import { PageBreadcrumbProps } from './PageBreadcrumb';
 import { PageGroup, PageGroupProps } from './PageGroup';
 import { getResizeObserver } from '../../helpers/resizeObserver';
 import { formatBreakpointMods, getBreakpoint, getVerticalBreakpoint } from '../../helpers/util';
+import { PageContextProvider } from './PageContext';
 
 export enum PageLayouts {
   vertical = 'vertical',
   horizontal = 'horizontal'
 }
-
-export interface PageContextProps {
-  isManagedSidebar: boolean;
-  onNavToggle: () => void;
-  isNavOpen: boolean;
-  width: number;
-  height: number;
-  getBreakpoint: (width: number | null) => 'default' | 'sm' | 'md' | 'lg' | 'xl' | '2xl';
-  getVerticalBreakpoint: (height: number | null) => 'default' | 'sm' | 'md' | 'lg' | 'xl' | '2xl';
-}
-
-export const pageContextDefaults: PageContextProps = {
-  isManagedSidebar: false,
-  isNavOpen: false,
-  onNavToggle: () => null,
-  width: null,
-  height: null,
-  getBreakpoint,
-  getVerticalBreakpoint
-};
-export const PageContext = React.createContext<PageContextProps>(pageContextDefaults);
-
-export const PageContextProvider = PageContext.Provider;
-export const PageContextConsumer = PageContext.Consumer;
-
 export interface PageProps extends React.HTMLProps<HTMLDivElement> {
   /** Content rendered inside the main section of the page layout (e.g. <PageSection />) */
   children?: React.ReactNode;
